Fixes #362:
For certain datasets (i.e. Homesite) atomiAdd(float*, float) was *very* unstable.
This is due to the fact that, because of rounding errors, it is not a transitive operation
and we are computing new centroids using kernels which run in a parallel (undeterministic) fashion.

This fix uses Kahan summation, which works well but bumps the memory footprint and is here as a
'quick' fix. In the long term we can change the centroids matrix to always be of type double.
This change will require a bit more time, though.